+++ Model/Config/Settings.php
@@ -9,6 +9,7 @@ class Settings
 {
     public const GOOGLE_AUTOCOMPLETE_API_KEY_XML_PATH = 'google_autocomplete/general/api_key';
     const CUSTOMER_STREET_LINES_QTY_CONFIG_PATH = 'customer/address/street_lines';
+    const DEFAULT_COUNTRY_CONFIG_PATH = 'general/country/default';
 
     /**
      * @var ScopeConfigInterface
@@ -31,6 +32,14 @@ class Settings
     {
         return (string) $this->scopeConfig->getValue(
             self::GOOGLE_AUTOCOMPLETE_API_KEY_XML_PATH,
+            ScopeInterface::SCOPE_STORE
+        );
+    }
+
+    public function getDefaultCountryId(): ?string
+    {
+        return $this->scopeConfig->getValue(
+            self::DEFAULT_COUNTRY_CONFIG_PATH,
             ScopeInterface::SCOPE_STORE
         );
     }