+++ ViewModel/Settings.php
@@ -24,6 +24,11 @@ class Settings implements ArgumentInterface
         return $this->settings->getApiKey();
     }
 
+    public function getDefaultCountryId(): ?string
+    {
+        return $this->settings->getDefaultCountryId();
+    }
+
     public function getStreetLinesQty(): string
     {
         return $this->settings->getStreetLinesQty();