+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 04-01-2022
+### Changed
+- [MAX-720](https://app.clickup.com/t/1y7qxvb) Set region autocomplete based on the default country id
+
 ## 1.1.1 - 28-12-2021
 ### Fixed
 - [MAX-721](https://app.clickup.com/t/1y7qzta) Cleaned up housenumber addition field if address doesn't have housenumber addition